refactor(file): Domain 형식으로 구조 개편
