+++ src/main/java/com/yeohangttukttak/api/domain/place/entity/Place.java
@@ -2,7 +2,7 @@ package com.yeohangttukttak.api.domain.place.entity;
 
 import com.yeohangttukttak.api.domain.Attachable;
 import com.yeohangttukttak.api.domain.BaseEntity;
-import com.yeohangttukttak.api.domain.file.File;
+import com.yeohangttukttak.api.domain.file.entity.File;
 import com.yeohangttukttak.api.domain.visit.entity.Visit;
 import jakarta.persistence.*;
 import lombok.Builder;
